Add IMDB support
Add support to search IMDB in the front end app if a RapidAPI key is provided to watchlist-backend  which does the search on IMDB
Add AUTH_KEY which the front end app must provide to the backend service in order for any API calls from the front end app to work. added as security measure.